pyproject.toml
[project]
name = "md-telegraph"
version = "0.1.0"
description = "Telegram bot that publishes Markdown to Telegraph"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiogram>=3.27.0",
    "bleach>=6.3.0",
    "httpx>=0.28.1",
    "markdown-it-py>=4.0.0",
    "pydantic-settings>=2.13.1",
    "telegraph>=2.2.0",
]

[project.scripts]
md-telegraph-bot = "app.main:main"

[dependency-groups]
dev = [
    "mypy>=1.20.0",
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.9",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B"]

[tool.pytest.ini_options]
pythonpath = ["."]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true
disable_error_code = ["import-untyped", "call-arg"]
